+++ 04. C# OOP Basics - June 2017/01. Defining Classes/Exercises/_009/_009/Rectangle.cs
@@ -1,10 +1,8 @@
 ﻿namespace _009
 {
-    using System;
-
     public class Rectangle
     {
-        public Rectangle(string id, int width, int height, int x, int y)
+        public Rectangle(string id, double width, double height, double x, double y)
         {
             this.ID = id;
             this.Width = width;
@@ -15,13 +13,13 @@
 
         public string ID { get; set; }
 
-        public int Width { get; set; }
+        public double Width { get; set; }
 
-        public int Height { get; set; }
+        public double Height { get; set; }
 
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         public static bool Check(Rectangle one, Rectangle two)
         {
